requirements.txt
vertexai
pydantic
instructor
openai
python-dotenv
logging
PyYAML
